requirements.txt
flask
flask-cors
pillow
numpy
face-recognition
gunicorn
